extract pooling module, tighten dispatcher, preserve 1_Pooling in convert
- Move pool helpers + dispatcher from `models/base.py` to a new
  `models/pooling.py`. `base.py` goes back to its pre-PR shape (data
  classes + `normalize_embeddings`).
- Update imports across all model files that use `mean_pooling`:
  `bert`, `xlm_roberta`, `modernbert`, `gemma3_text`, `llama_bidirec`,
  `lfm2`, `llama_nemotron_vl`.
- Tighten `pool_by_config` / `_normalize_pooling_config` signatures from
  `Optional[Dict[str, Any]]` to `Dict[str, Any]`; remove the now-unreachable
  `if cfg is None` fallback (the dataclass default factory guarantees a
  dict reaches the dispatcher).
- Tighten `bert.ModelArgs.pooling_config` and `xlm_roberta.ModelArgs.pooling_config`
  from `Optional[dict]` to `dict`.
- `convert.py`: preserve `1_Pooling/` subdirectory when converting an HF
  checkpoint. The top-level `*.json` glob doesn't recurse; without this,
  converted `mlx-community/*` variants lose the pooling sidecar and the
  loader silently falls back to mean.
- Split `tests/test_base.py` -> `tests/test_pooling.py` for the new
  pool-helper / dispatcher / config-normalization tests; replace
  per-test "Port of test_xxx" annotations with the upstream comments
  verbatim and add the two HF-port dispatcher tests
  (`test_forward_all_modes`, `test_invalid_mode_raises`).
