+++ mlx_embeddings/models/bert.py
@@ -5,7 +5,8 @@ from typing import Optional, Tuple
 import mlx.core as mx
 import mlx.nn as nn
 
-from .base import BaseModelArgs, BaseModelOutput, normalize_embeddings, pool_by_config
+from .base import BaseModelArgs, BaseModelOutput, normalize_embeddings
+from .pooling import pool_by_config
 
 
 @dataclass
@@ -22,7 +23,7 @@ class ModelArgs(BaseModelArgs):
     initializer_range: float = 0.02
     layer_norm_eps: float = 1e-12
     vocab_size: int = 30522
-    pooling_config: Optional[dict] = field(
+    pooling_config: dict = field(
         default_factory=lambda: {"pooling_mode": "mean"}
     )
 
@@ -227,11 +228,6 @@ class Model(nn.Module):
         sequence_output = encoder_outputs
         pooled_output = self.pooler(sequence_output)
 
-        # Sentence-transformers-compatible pooling: dispatch on the
-        # `1_Pooling/config.json` mode (cls / mean / max / mean_sqrt_len)
-        # surfaced by the loader. Falls back to mean pooling when no pooling
-        # config is present, preserving prior behaviour for plain HF
-        # checkpoints.
         text_embeds = pool_by_config(
             sequence_output, attention_mask, self.config.pooling_config
         )
